+++ spec/logging_spec.rb
@@ -74,7 +74,7 @@ describe 'openstack-common::logging' do
             ['[loggers]',
              'keys=root,ceilometer,cinder,glance,horizon,keystone,nova,'\
                'neutron,swift,trove,amqplib,sqlalchemy,boto,suds,eventletwsgi,'\
-               'nova_api_openstack_wsgi,nova_osapi_compute_wsgi_server,'],
+               'nova_api_openstack_wsgi,nova_osapi_compute_wsgi_server'],
             ['[logger_root]',
              'level=NOTSET',
              'handlers=devel'],